première version line chart
