+++ app/src/main/java/com/example/b_rich/MainActivity.kt
@@ -1,15 +1,12 @@
 package com.example.b_rich
 
-import ExchangeRate
 import android.annotation.SuppressLint
 import android.net.Uri
 import android.os.Bundle
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.DrawerValue
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.rememberDrawerState
 import androidx.compose.ui.Modifier
 import androidx.fragment.app.FragmentActivity
 import androidx.navigation.NavController
@@ -20,13 +17,14 @@ import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navArgument
 import com.example.b_rich.data.entities.user
 import com.example.b_rich.data.network.RetrofitClient
+import com.example.b_rich.data.repositories.CurrencyConverterRepository
 import com.example.b_rich.data.repositories.ExchangeRateRepository
 import com.example.b_rich.data.repositories.UserRepository
-import com.example.b_rich.ui.AddAccount.AddAccountScreen
 import com.example.b_rich.ui.AddAccount.AddAccountViewModel
 import com.example.b_rich.ui.MainScreen
+import com.example.b_rich.ui.currency_converter.CurrencyConverter
+import com.example.b_rich.ui.currency_converter.CurrencyConverterViewModel
 import com.example.b_rich.ui.exchange_rate.ExchangeRateViewModel
-import com.example.b_rich.ui.profil.UserProfileScreen
 import com.example.b_rich.ui.resetPassword.CodeEntryScreen
 import com.example.b_rich.ui.resetPassword.PasswordEntryScreen
 import com.example.b_rich.ui.resetPassword.ResetPasswordViewModel
@@ -35,11 +33,9 @@ import com.example.b_rich.ui.signin.SigninViewModel
 import com.example.b_rich.ui.signup.SignUpScreen
 import com.example.b_rich.ui.signup.SignupViewModel
 import com.example.b_rich.ui.theme.BrichTheme
-import com.example.b_rich.ui.wallets.HomeBrichScreen
 import com.example.b_rich.ui.wallets.HomeViewModel
 import com.example.b_rich.ui.welcome.WelcomeScreen
 import com.google.gson.Gson
-import kotlinx.coroutines.flow.toList
 
 class MainActivity : FragmentActivity() {
     @SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
@@ -51,11 +47,14 @@ class MainActivity : FragmentActivity() {
         val apiService = RetrofitClient.getApiService()
         val userRepository = UserRepository(apiService)
         val exchangeRateRepository =ExchangeRateRepository(apiService)
+        val currencyRepository= CurrencyConverterRepository(apiService)
+
         val signinViewModel = SigninViewModel(userRepository)
         val resetPasswordViewModel = ResetPasswordViewModel(userRepository)
         val signupViewModel= SignupViewModel(userRepository)
         val exchangeRateViewModel=ExchangeRateViewModel(exchangeRateRepository)
         val addAccountViewModel= AddAccountViewModel()
+        val currencyConverterViewModel=CurrencyConverterViewModel(currencyRepository)
         val HomeViewModel=HomeViewModel()
 
 
@@ -74,21 +73,23 @@ class MainActivity : FragmentActivity() {
                                wallets = HomeViewModel.wallets
                                )
                         }*/
-                        composable("Comptes_bancaires") {
+                        /*composable("Comptes_bancaires") {
                             val drawerState = rememberDrawerState(initialValue = DrawerValue.Closed)
                             AddAccountScreen(
                                 navHostController = navController,
                                 drawerState = drawerState,
                                 viewModel = addAccountViewModel
                             )
-                        }
+                        }*/
                         composable("signup"){
                             SignUpScreen(signupViewModel,navController)
                         }
                         composable("loginPage") {
                             LoginScreen(signinViewModel, navController)
                         }
-
+                        composable("currencyConvert") {
+                            CurrencyConverter(currencyConverterViewModel)
+                        }
                         composable(
                             route = "exchangeRate/{userJson}",
                             arguments = listOf(
@@ -97,7 +98,7 @@ class MainActivity : FragmentActivity() {
                         ) { backStackEntry ->
                             val userJson = backStackEntry.arguments?.getString("userJson")
                             val user = userJson?.let { Gson().fromJson(it, user::class.java) }
-                            user?.let { MainScreen(it, navController, resetPasswordViewModel,exchangeRateViewModel,addAccountViewModel) }
+                            user?.let { MainScreen(it, navController, resetPasswordViewModel,exchangeRateViewModel,addAccountViewModel,currencyConverterViewModel) }
                         }
 
                         composable(
