+++ app/src/main/java/com/example/b_rich/data/dataModel/PredictionData.kt
@@ -1,0 +1,6 @@
+package com.example.b_rich.data.dataModel
+
+data class PredictionData(
+    val date: String,
+    val value: Double
+)
