+++ app/src/main/java/com/example/b_rich/data/dataModel/PredictionRequest.kt
@@ -1,0 +1,6 @@
+package com.example.b_rich.data.dataModel
+
+data class PredictionRequest(
+    val date: String,
+    val currencies: List<String>
+)
