+++ app/src/main/java/com/example/b_rich/data/dataModel/PredictionResponse.kt
@@ -1,0 +1,6 @@
+package com.example.b_rich.data.dataModel
+
+data class PredictionResponse(
+    val predictions: Map<String, List<PredictionData>>,
+    val start_date: String
+)
