+++ app/src/main/java/com/example/b_rich/data/network/ApiService.kt
@@ -1,5 +1,6 @@
 package com.example.b_rich.data.network
 
+import com.example.b_rich.data.dataModel.PredictionResponse
 import com.example.b_rich.data.entities.ExchangeRate
 import com.example.b_rich.data.entities.user
 import retrofit2.Response
@@ -77,4 +78,9 @@ interface ApiService {
         @Path("amount") amount: String
     ): Double
 
+    @POST("predictions/create-prediction")
+    suspend fun getPredictions(
+        @Body date: String,
+        @Body currencies: List<String>
+    ): Response<PredictionResponse>
 }
