+++ app/src/main/java/com/example/b_rich/data/network/RetrofitClient.kt
@@ -6,7 +6,7 @@ import retrofit2.converter.gson.GsonConverterFactory
 
 object RetrofitClient {
 
-    private const val BASE_URL = "http://192.168.1.102:3000/"
+    private const val BASE_URL = "http://192.168.88.164:3000/"
 
     private val retrofit by lazy {
         val client = OkHttpClient.Builder().build()
